backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.31
aiosqlite==0.20.0
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.3.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
duckdb==0.10.3
pdfplumber==0.11.2
sentence-transformers==2.7.0
pandas==2.2.2
langchain-core==0.2.38
langchain-groq==0.1.9
langchain-anthropic==0.1.23
httpx==0.27.0
pytest==8.3.2
pytest-asyncio==0.23.8
